sensor/requirements.txt
influxdb
paho-mqtt
rpi.gpio
RPi.bme280